refactor(phase-2-3): Timerコンポーネントの分割
## 変更内容
- Timer.tsx (352行) を5つのサブコンポーネントに分割
  - index.tsx: メインコンポーネント (143行, 59%削減)
  - TimerDisplay.tsx: 時間表示 (52行)
  - TimerControls.tsx: 操作ボタン (80行)
  - TimerProjectInfo.tsx: プロジェクト情報表示 (78行)
  - TimerBackground.tsx: 背景アニメーション (59行)
- 各サブコンポーネントのテストを追加 (13テスト)
- 既存テスト全て通過 (321テスト)

## Phase 2 完了
- P2-1: ProjectProgressCard分割 ✅
- P2-2: useStorage分割 ✅
- P2-3: Timer整理 ✅
