+++ docs/refactoring/phased-refactoring-plan-2026.md
@@ -175,10 +175,10 @@ export * from './themeHelpers';
 
 ---
 
-## Phase 2: コンポーネント分割（アーキテクチャ改善） 🚧 進行中
+## Phase 2: コンポーネント分割（アーキテクチャ改善） ✅ 完了
 
 **目的**: 単一責任原則に基づくコンポーネント設計
-**ステータス**: P2-1完了、P2-2完了、P2-3未着手
+**ステータス**: P2-1完了、P2-2完了、P2-3完了
 
 ### P2-1: ProjectProgressCard のリファクタリング ✅
 
@@ -293,21 +293,33 @@ src/services/
 - [x] 単一責任原則を遵守
 - [x] 新規ファイルにテストを追加（24テスト追加）
 
-### P2-3: Timer コンポーネントの整理
+### P2-3: Timer コンポーネントの整理 ✅
 
-**現状**: 352行
+**結果**: 352行 → 143行（59%削減）
 
-**目標構成**:
+**実現構成**:
 
 ```
 src/components/timer/
-├── Timer/
-│   ├── index.tsx
-│   ├── TimerDisplay.tsx      # 時間表示
-│   ├── TimerControls.tsx     # 操作ボタン
-│   ├── ProjectSelector.tsx   # プロジェクト選択
-│   └── __tests__/
+├── Timer.tsx                 # 再エクスポート (1行)
+└── Timer/
+    ├── index.tsx             # メインコンポーネント (143行)
+    ├── TimerDisplay.tsx      # 時間表示 (52行)
+    ├── TimerControls.tsx     # 操作ボタン (80行)
+    ├── TimerProjectInfo.tsx  # プロジェクト情報表示 (78行)
+    ├── TimerBackground.tsx   # 背景アニメーション (59行)
+    └── __tests__/
+        ├── TimerDisplay.test.tsx      # 4テスト
+        ├── TimerControls.test.tsx     # 5テスト
+        └── TimerProjectInfo.test.tsx  # 4テスト
 ```
+
+**完了条件**:
+
+- [x] コンポーネントが適切に分割
+- [x] 単一責任原則を遵守
+- [x] 新規ファイルにテストを追加（13テスト追加）
+- [x] 既存テストが全て通る（321テスト）
 
 ---
